requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Model architectures
timm>=0.9.0

# Data processing and augmentation
albumentations>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Jupyter (optional, for EDA)
jupyter>=1.0.0
ipykernel>=6.25.0
